requirements.txt
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
dnspython==2.2.1
exceptiongroup==1.1.0
fastapi==0.88.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==1.1.1
mongomock==4.1.2
motor==3.1.1
packaging==22.0
pluggy==1.0.0
pydantic==1.10.4
pymongo==4.3.3
pytest==7.2.0
python-dotenv==0.21.0
rfc3986==1.5.0
sentinels==1.0.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
